requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
pydub==0.25.1
openai==1.35.13
python-dotenv==1.0.1
-e pythonlib
